+++ src/serialization/__serialize_267.cpp
@@ -28,8 +28,8 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<RegressionThresholdOptimizer<
-              float,
-              float
+              double,
+              double
 	  >>();
     }
   };
@@ -43,16 +43,16 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const RegressionThresholdOptimizer<
-              float,
-              float
+              double,
+              double
 	  > *, const bool &);
     TemplateFuncExport DllExport RegressionThresholdOptimizer<
-              float,
-              float
+              double,
+              double
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, RegressionThresholdOptimizer<
-              float,
-              float
+              double,
+              double
 	  >*);
 }  // namespace fertilized
 
@@ -63,8 +63,8 @@ namespace serialization {
 
 template <>
 struct version<RegressionThresholdOptimizer<
-              float,
-              float
+              double,
+              double
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
     typedef mpl::integral_c_tag tag;
