+++ src/serialization/__serialize_299.cpp
@@ -28,9 +28,9 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<Tree<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >>();
@@ -46,23 +46,23 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const Tree<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  > *, const bool &);
     TemplateFuncExport DllExport Tree<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, Tree<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >*);
@@ -75,9 +75,9 @@ namespace serialization {
 
 template <>
 struct version<Tree<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >> {
